feat: create separate fixtures
